### 2020-06-08 a35dd23
Return identifier from `CommunicatorHelper.add_rpc_subscriber` (#64)
The spec of the `Communicator` abstract class stipulates that the
`add_rpc_subscriber` should always return the identifier, even if an
explicit one was passed in. The `CommunicatorHelper` violated this
specification but went unnoticed since it was not tested.
