+++ kiwipy/version.py
@@ -1,3 +1,3 @@
-__version__ = "0.5.4"
+__version__ = "0.5.5"
 
 __all__ = ['__version__']
