### 2020-06-16 #2295 @danielskatz
👋 @muhrin - if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).  I will also think about potential reviewers...
