### 2020-06-16 #2295 @danielskatz
Also, the repository you entered when submitting (as shown at the top of the issue is https://github.com/aiidateam/kiwipy.git - I'm going to change this to https://github.com/aiidateam/kiwipy unless you object...)
