### 2020-06-17 #2295 @uellue
> 👋 @uellue - would you be willing to review this submission for JOSS?

Yes, sure! :-)
