### 2020-06-17 #2351 @whedon
[REVIEW]: kiwiPy: Robust, high-volume, messaging for big-data and computational science workflows
**Submitting author:** @muhrin (<a href="http://orcid.org/0000-0001-6902-1289">Martin Uhrin</a>)
**Repository:** <a href="https://github.com/aiidateam/kiwipy" target ="_blank">https://github.com/aiidateam/kiwipy</a>
**Version:** v0.6.1
