### 2020-06-19 aiidateam/kiwipy#65 @uellue
Add contributor instructions
From https://github.com/openjournals/joss-reviews/issues/2351:

Adding at least a short statement to the documentation on how to contribute would help with checking off this item in the review!
