### 2020-06-19 #2351 @uellue
Regarding "State of the field":

## Comparison with Pika

After a quick side-by-side comparison between the RabbitMQ documentation and kiwiPy, it is not clear to me what the difference between the default Pika Python API and the kiwiPy API is. The code of the Pika-based RabbitMQ examples and the kiwiPy examles are quite similar, and at least to my eyes the kiwiPy version is not a noticeable improvement, different from what is claimed in the paper.

<details><summary>Pika https://www.rabbitmq.com/tutorials/tutorial-two-python.html</summary>
#!/usr/bin/env python
import time

import pika

connection = pika.BlockingConnection(
    pika.ConnectionParameters(host='localhost'))
channel = connection.channel()

channel.queue_declare(queue='hello')

def callback(ch, method, properties, body):
    print(" [x] Received %r" % body)
    time.sleep(body.count(b'.'))
    print(" [x] Done")

channel.basic_consume(
    queue='hello', on_message_callback=callback, auto_ack=True)

print(' [*] Waiting for messages. To exit press CTRL+C')
channel.start_consuming()

<details><summary>kiwiPy https://github.com/aiidateam/kiwipy/blob/develop/README.rst</summary>
# -*- coding: utf-8 -*-
import time
import threading

import kiwipy

print(' [*] Waiting for messages. To exit press CTRL+C')

def callback(_comm, task):
    print((' [x] Received %r' % task))
    time.sleep(task.count('.'))
    print(' [x] Done')
    return task

with kiwipy.connect('amqp://127.0.0.1/') as communicator:
    communicator.add_task_subscriber(callback)
    try:
        threading.Event().wait()
    except KeyboardInterrupt:
        pass

## Comparison with other work distribution and messaging systems

A quick list of major distributed work scheduling systems with Python API that might deserve a comparison:

* Dask futures https://docs.dask.org/en/latest/futures.html
* MPI
* ZeroMQ
* An endless list of distributed task queue systems: https://www.fullstackpython.com/task-queues.html
* An endless list of Python-based RPC systems: https://stackoverflow.com/questions/1879971/what-is-the-current-choice-for-doing-rpc-in-python

The paper and documentation should make much clearer what makes kiwiPy unique and for what particular use cases it is advantageous. Perhaps it should reference the appropriate parts of RabbitMQ documentation since it is essentially a Python API for it, as far as I understood.
