### 2020-06-19 #2351 @uellue
Regarding "Community guidelines": https://github.com/aiidateam/kiwipy/issues/65
