### 2020-06-19 aiidateam/kiwipy#65 @danielskatz
from https://github.com/openjournals/joss-reviews/issues/2351
