### 2020-06-23 53dc412
Added note to documentation on how to contribute
Added box on index page of docs pointing readers to the wiki which
explains how they can contribute to kiwiPy.
