### 2020-06-23 26b561a
Add note to documentation and README on how to contribute (#66)
Added box on index page of docs pointing readers to the wiki which
explains how they can contribute to kiwiPy.
