### 2020-06-26 #2351 @muhrin
We've addressed @uellue issue regarding contributor guidelines by adding [wiki entry](https://github.com/aiidateam/kiwipy/wiki/Contributing) with is now linked to in the docs.
I've yet to find the time to address the other points raise but should find the time in the coming days.  I'll report back as soon as I can.
