### 2020-07-01 ede53d7
Added comparison with pika to docs
As requested by our JOSS submission review I've added a comparison on
the differences to pika which also acts as a nice transition guide if
someone wants to come to kiwi from pika.
