### 2020-07-01 16b9128
Added reference to pika comparison in paper
Also excluded encoding pragma from code snippets in documentation.
