### 2020-07-01 c04c201
Added reference to pika comparison in paper
Also excluded encoding pragma from code snippets in documentation.
