+++ docs/source/examples/pika/rpc_server.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import threading
 import kiwipy
 
