+++ docs/source/examples/pika/receive_logs_topic.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import kiwipy
 import sys
 import threading
