+++ docs/source/examples/pika/new_task.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import kiwipy
 import sys
 
