+++ docs/source/examples/pika/worker.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import threading
 import kiwipy
 import time
