### 2020-07-01 3f68051
Added comparison to other libraries
This has helped to define better exactly the features and uses cases
that kiwiPy is best suited for.

Added comparisons to:
* pika
* aio-pika
* py-amqp
* kombu
* celery
* rq
* RPyC
* Spyne
* Python-JRPC
* Dask
* MPI
* ZeroMQ
