+++ docs/source/paper.md
@@ -28,16 +28,16 @@ bibliography: paper.bib
 
 The computational sciences have seen a huge increase in the use of high-throughput, automated, workflows over the course of the last two decades or so.  Focusing on just our domain of computational materials science, there have been several large scale initiatives to provide high-quality results from standardised calculations [@Landis2012; @Curtarolo2012; @Jain2013; @Saal2013; @Draxl2019a; @Talirz2020].  Almost all of these repositories are populated using results from high-throughput quantum mechanical calculations that rely on workflow frameworks [@Jain2015a; @Mortensen2020], including our own [AiiDA](http://www.aiida.net/) [@Pizzi2016; @Huber2020] which powers the [Materials Cloud](https://www.materialscloud.org/).  One of the many challenges for such frameworks is maximising fault-tolerance whilst simultaneously maintaining high-throughput, often across several systems (typically the client launching the tasks, the supercomputer carrying out the computations and the server hosting the database).
 
-On the software level, these problems are perhaps best addressed by using messaging brokers that take responsibility for guaranteeing the durability and atomicity of messages and often enable event-based communication.
+On the software level, these problems are perhaps best addressed by using messaging brokers that take responsibility for guaranteeing the durability (or persistence) and atomicity of messages and often enable event-based communication.
 Indeed, solutions such as [RabbitMQ](https://www.rabbitmq.com/), see widespread adoption in industry, however, adoption in academia has been more limited, with home-made queue data structures, race condition susceptible locks and polling based solutions being commonplace.
 This is likely due to message brokers typically having complex APIs (which reflect the non-trivial nature of the underlying protocol) as well as the lack of familiarity with event-based programming in general within the community.
-[KiwiPy](https://kiwipy.readthedocs.io/en/latest/) was designed specifically to address both these issues, by providing a tool that allows building robust, event-based systems with an interface that is as simple as possible.
+[KiwiPy](https://kiwipy.readthedocs.io/en/latest/) was designed specifically to address both these issues, by providing a tool that enables building robust, event-based systems with an interface that is as simple as possible.
 
-KiwiPy provides three main message types to the user: task queues, Remote Procedure Calls (RPCs), and, broadcasts.
-These are all exposed via one class called the `Communicator` which can be trivially constructed by providing a URI string pointing to the RabbitMQ server.
-This is in contrast with the several steps that are required to establish a connection and perform each of the communication types listed above using [Pika](https://pika.readthedocs.io/en/stable/), the standard library used in RabbitMQ [tutorials](https://www.rabbitmq.com/getstarted.html).
- A more in depth of the two can be found in our [documentation](https://kiwipy.readthedocs.io/en/latest/).
-Furthermore, by default, kiwiPy creates a separate communication thread that the user never sees, allowing them to interact with the communicator using familiar Python syntax, without the need to be familiar with either coroutines or multithreading.  This has the additional advantage that kiwiPy will maintain heartbeats (a periodic check to make sure the connection is still alive) with the server whilst the user code can be doing other things. Heartbeats are an essential part of RabbitMQ's fault tolerance whereby two missed checks will automatically trigger the message to be requeued to be picked up by another client.
[…]
