+++ docs/source/examples/pika/rpc_client.py
@@ -1,6 +1,42 @@
-import kiwipy
+import pika
+import uuid
 
-with kiwipy.connect('amqp://localhost') as comm:
-    print(' [x] Requesting fib(30)')
-    response = comm.rpc_send('rpc_queue', 30).result()
-    print(' [.] Got %r' % response)
+
+class FibonacciRpcClient(object):
+
+    def __init__(self):
+        self.connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost'))
+
+        self.channel = self.connection.channel()
+
+        result = self.channel.queue_declare(queue='', exclusive=True)
+        self.callback_queue = result.method.queue
+
+        self.channel.basic_consume(queue=self.callback_queue, on_message_callback=self.on_response, auto_ack=True)
+
+    def on_response(self, ch, method, props, body):
+        if self.corr_id == props.correlation_id:
+            self.response = body
+
+    def call(self, n):
+        self.response = None
+        self.corr_id = str(uuid.uuid4())
+        self.channel.basic_publish(
+            exchange='',
+            routing_key='rpc_queue',
+            properties=pika.BasicProperties(
+                reply_to=self.callback_queue,
+                correlation_id=self.corr_id,
+            ),
+            body=str(n)
+        )
+        while self.response is None:
+            self.connection.process_data_events()
+        return int(self.response)
+
+
+fibonacci_rpc = FibonacciRpcClient()
+
+print(' [x] Requesting fib(30)')
+response = fibonacci_rpc.call(30)
+print(' [.] Got %r' % response)
