+++ docs/source/examples/pika/rpc_server.py
@@ -1,5 +1,10 @@
-import threading
-import kiwipy
+import pika
+
+connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost'))
+
+channel = connection.channel()
+
+channel.queue_declare(queue='rpc_queue')
 
 
 def fib(n):
@@ -11,14 +16,22 @@ def fib(n):
         return fib(n - 1) + fib(n - 2)
 
 
-def on_request(_comm, body):
+def on_request(ch, method, props, body):
     n = int(body)
 
     print(' [.] fib(%s)' % n)
-    return fib(n)
+    response = fib(n)
+
+    ch.basic_publish(exchange='',
+                     routing_key=props.reply_to,
+                     properties=pika.BasicProperties(correlation_id= \
+                                                         props.correlation_id),
+                     body=str(response))
+    ch.basic_ack(delivery_tag=method.delivery_tag)
 
 
-with kiwipy.connect('amqp://localhost') as comm:
-    comm.add_rpc_subscriber(on_request, identifier='rpc_queue')
-    print(' [x] Awaiting RPC requests')
-    threading.Event().wait()
+channel.basic_qos(prefetch_count=1)
+channel.basic_consume(queue='rpc_queue', on_message_callback=on_request)
+
+print(' [x] Awaiting RPC requests')
+channel.start_consuming()
