+++ docs/source/examples/pika/receive_logs_topic.py
@@ -1,20 +1,29 @@
-import kiwipy
+import pika
 import sys
-import threading
+
+connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost'))
+channel = connection.channel()
+
+channel.exchange_declare(exchange='topic_logs', exchange_type='topic')
+
+result = channel.queue_declare('', exclusive=True)
+queue_name = result.method.queue
+
+binding_keys = sys.argv[1:]
+if not binding_keys:
+    sys.stderr.write('Usage: %s [binding_key]...\n' % sys.argv[0])
+    sys.exit(1)
+
+for binding_key in binding_keys:
+    channel.queue_bind(exchange='topic_logs', queue=queue_name, routing_key=binding_key)
+
+print(' [*] Waiting for logs. To exit press CTRL+C')
 
 
-def callback(_comm, body, _sender, subject, _msg_id):
-    print(' [x] %r:%r' % (subject, body))
+def callback(ch, method, properties, body):
+    print(' [x] %r:%r' % (method.routing_key, body))
 
 
-with kiwipy.connect('amqp://localhost') as comm:
-    binding_keys = sys.argv[1:]
-    if not binding_keys:
-        sys.stderr.write('Usage: %s [binding_key]...\n' % sys.argv[0])
-        sys.exit(1)
+channel.basic_consume(queue=queue_name, on_message_callback=callback, auto_ack=True)
 
-    for binding_key in binding_keys:
-        comm.add_broadcast_subscriber(kiwipy.BroadcastFilter(callback, binding_key))
-
-    print(' [*] Waiting for logs. To exit press CTRL+C')
-    threading.Event().wait()
+channel.start_consuming()
