+++ docs/source/examples/pika/emit_log_topic.py
@@ -1,9 +1,13 @@
-import kiwipy
+import pika
 import sys
 
-message = ' '.join(sys.argv[2:]) or 'Hello World!'
+connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost'))
+channel = connection.channel()
 
-with kiwipy.connect('amqp://localhost') as comm:
-    subject = sys.argv[1] if len(sys.argv) > 2 else 'anonymous.info'
-    comm.broadcast_send(message, subject=subject)
-    print(' [x] Sent %r:%r' % (subject, message))
+channel.exchange_declare(exchange='topic_logs', exchange_type='topic')
+
+routing_key = sys.argv[1] if len(sys.argv) > 2 else 'anonymous.info'
+message = ' '.join(sys.argv[2:]) or 'Hello World!'
+channel.basic_publish(exchange='topic_logs', routing_key=routing_key, body=message)
+print(' [x] Sent %r:%r' % (routing_key, message))
+connection.close()

[…]
