### 2020-07-06 384bfc0
fix 70: wrap kiwipy future threadsafe
