### 2020-07-06 56a6050
Docs: add comparison to `pika` with examples (#68)
As requested by our JOSS submission review add a comparison on
the differences to `pika` which also acts as a nice transition guide if
someone wants to come to `kiwipy` from `pika`.
