### 2020-07-07 22379e8
Update docs/source/index.rst
Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
