### 2020-07-07 92fea54
Update docs/source/paper.md
Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
