+++ docs/source/paper.md
@@ -33,7 +33,7 @@ Indeed, solutions such as [RabbitMQ](https://www.rabbitmq.com/), see widespread 
 This is likely due to message brokers typically having complex APIs (which reflect the non-trivial nature of the underlying protocol) as well as the lack of familiarity with event-based programming in general within the community.
 [KiwiPy](https://kiwipy.readthedocs.io/en/latest/) was designed specifically to address both these issues, by providing a tool that enables building robust, event-based systems with an interface that is as simple as possible.
 
-In kiwiPy all messages are saved to disk by RabbitMQ meaning that any or all systems involved in a workflow, including the broker, can be shut down (abruptly or gracefully), and the previous state can be recreated allowing the workflow to continue when the necessary resources are brought back online.  This is especially important for long-running HPC jobs which may take days or weeks.  This feature differentiates kiwiPy from protocols such as MPI, [ZeroMQ](https://zeromq.org/) or libraries such as [Dask](https://docs.dask.org/en/latest/) which do not persist their state.
+In kiwiPy, all messages are saved to disk by RabbitMQ, meaning that any or all systems involved in a workflow, including the broker, can be shut down (abruptly or gracefully), and the previous state can be recreated allowing the workflow to continue when the necessary resources are brought back online.  This is especially important for long-running HPC jobs which may take days or weeks.  This feature differentiates kiwiPy from protocols such as MPI, [ZeroMQ](https://zeromq.org/) or libraries such as [Dask](https://docs.dask.org/en/latest/) which do not persist their state.
 
 There exist a number of libraries for interacting directly with message brokers including [Pika](https://pika.readthedocs.io/en/stable/), [aio-pika](https://aio-pika.readthedocs.io/en/latest/index.html), [py-amqp](https://barryp.org/software/py-amqplib), [kombu](https://github.com/celery/kombu) and others.  These tend to be rather low-level and focus on channels, exchanges, routing, sockets and so on.   A comparison of the difference in focus between Pika and kiwiPy can be found in the [documentation](https://kiwipy.readthedocs.io/en/latest/).  On the other end of the spectrum exist libraries such as [Celery](https://docs.celeryproject.org/en/stable/getting-started/introduction.html), [RQ](https://python-rq.org/) and [others](https://www.fullstackpython.com/task-queues.html) that provide task queues and libraries such as [RPyC](https://rpyc.readthedocs.io/en/latest/), [Spyne](http://spyne.io), [Python-JRPC](https://github.com/alex-sherman/python-jrpc) and [others](https://stackoverflow.com/questions/1879971/what-is-the-current-choice-for-doing-rpc-in-python) that enable remote procedure calls.  In contrast, kiwiPy brings together three commonly used message types (task queues, Remote Procedure Calls (RPCs), and, broadcasts) in a single interface.
 
