### 2020-07-07 e10464d
Broader comparison to other similar libraries (#69)
* Added comparison to other libraries

This has helped to define better exactly the features and uses cases
that kiwiPy is best suited for.

Added comparisons to:
* pika
* aio-pika
* py-amqp
* kombu
* celery
* rq
* RPyC
* Spyne
* Python-JRPC
* Dask
* MPI
* ZeroMQ
