### 2020-07-07 #2351 @muhrin
Dear @danielskatz,

we once again thank @uellue for his thorough review which has helped to improve both the documentation and the paper itself.  We have made the following changes to address the comments made:

* https://github.com/aiidateam/kiwipy/commit/26b561ab32561e8962e6a1126d36f2072232daa0
We have added a [contributing guide](https://github.com/aiidateam/kiwipy/wiki/Contributing) to our wiki detailing how contributions to kiwiPy can be made (this is linked to from the [documentation](https://kiwipy.readthedocs.io/en/latest/index.html)), thus closing aiidateam/kiwipy#65.

* https://github.com/aiidateam/kiwipy/commit/56a6050509e3aedaf918a91eaa9de1b15cb088c6
 A [direct comparison](https://kiwipy.readthedocs.io/en/latest/pika-comparison.html) to `pika` (the library used for tutorial on the RabbitMQ website) has been added to the documentation that both highlights the differences in in syntax and acts as a guide for transitioning to kiwiPy from pika.  We believe this highlights that, for all but the most simple of cases, kiwiPy has more compact and pythonic syntax (e.g. contexts for resource lifetime scoping, use of futures), but perhaps more significantly we show that kiwiPy has a higher-level focus compared to pika making it more accessible to users not familiar with channels, exchanges, routing, etc.

* https://github.com/aiidateam/kiwipy/commit/e10464df91162ba8781cf205493246161eb5c9d1
We recognise that we did not sufficiently emphasise the major benefit and differentiating factor of kiwiPy as compared to some existing libraries as highlighted in the review (Dask, ZMQ, MPI).  Namely, that kiwiPy (via RabbitMQ) automatically persists messages (e.g. tasks) to disk allowing us to build loosely coupled systems where it is not expected that all clients (or even the broker) will be online at all times allowing long-running workflows (taking weeks or months) to complete as and when resources are available.  This has been highlighted in the paper and a comparison has been added to a number of existing libraries.

This concludes our changes in response to the review.

We look forward to the continuation of the review process.
