### 2020-07-07 #2351 @uellue
Dear @muhrin,

many thanks for the update! The new content is a great help for readers to understand the use cases and strengths of kiwiPy. The comparison with Pika and other systems sounds fair and objective. :+1: This fully addresses my previous comments and from my perspective the paper is good to go. 🚀 Congratulations!

CC @danielskatz
