### 2020-07-07 #2351 @dghoshal-lbl
I finished my first round of review for the software and the paper. Some of the paper related concerns have already been addressed. I have a couple of more questions regarding the paper:

i) Has the library been tested to show any significant performance overheads? Mainly because this adds another level of abstraction to RabbitMQ.

ii) Since one of the use-cases for kiwiPy is to restart partially completed workflows and HPC jobs, I was wondering if there are any specific configurations in kiwiPy that enables this or if this is a default behavior?

Regarding the software, I was able to install it but not test its functionality properly:

i) I could not find any module or scripts that would let me test the library.

ii) When I wrote my own client and server following the examples, I ran into the following error:
`ImportError: cannot import name 'AsyncContextManager'`

I also see that there are a few examples in the source tree. I can dig deeper into why I am getting this error, but I think it will be good to document the steps to quickly test the examples or provide scripts for validating the tests.
