### 2020-07-13 #2351 @muhrin
Thanks @danielskatz.

Thank @dhhagan for your review.  To answer your points:

1.i) I'm running some benchmarks comparing kiwiPy with pika (the 'default' python library for interacting with RabbitMQ).  I'll post back with results shortly.

1.ii) Yes, this is the default.  KiwiPy (in the language of RMQ) uses persistent messages and durable queues/exchanges so it's like this out of the box.  We tried to highlight this in the paper by stating:

_"In kiwiPy, all messages are saved to disk by RabbitMQ, meaning that any or all systems involved in a workflow, including the broker, can be shut down (abruptly or gracefully), and the previous state can be recreated allowing the workflow to continue when the necessary resources are brought back online."_

If you feel that this could benefit from some specific clarification we would be happy to add it.

In an HPC setting this persistence can be exploited by submitting a python script to the scheduler that works on one or more kiwiPy tasks.  If the scheduler job is terminated the running task is automatically re-inserted back into kiwiPy's queue and will be picked up the next time that job script is submitted.  We have deliberately tried to keep kiwiPy's documentation somewhat use-case agnostic as HPC is just one setting where it can be useful, however there is a detailed description in an upcoming [AiiDA](www.aiida.net) paper of a similar usage scenario.

2.i) As far as testing kiwiPy, thanks for pointing out that this isn't documented.  I will add it to the [installation](https://kiwipy.readthedocs.io/en/latest/installation.html) instructions shortly.  Briefly, for someone that has cloned the source three it is just a matter of:
```
pip install -e .[rmq,tests]
pytest test
```
2.ii) Regarding the `AsyncContextManager` exception, would you be so kind as to open an issue [here](https://github.com/aiidateam/kiwipy/issues) with your client/server code and I will have a look?
