### 2020-07-13 84f7588
Added benchmarks comparing to pika
