### 2020-07-13 8240587
Added instructions on how to run tests
Added to the installation section.
