### 2020-07-14 4b910a0
Added benchmarks comparing to pika
Added benchmarks to test suite as well as a page on the docs showing the difference in performance wrt pika.
A question about this was raised during our JOSS review: openjournals/joss-reviews#2351.
