### 2020-07-14 8b7cd5d
Add benchmark tests and docs comparing to pika (#73)
Added benchmarks to test suite as well as a page on the docs showing the
difference in performance with respect to `pika`. A question about this
was raised during our JOSS review: openjournals/joss-reviews#2351.
