### 2020-07-14 #2351 @muhrin
Dear @dghoshal-lbl ,

I'm writing back to share changes addressing the points you've raised.

* https://github.com/aiidateam/kiwipy/pull/72
Adds instructions on how to run the tests which can be seen live [here](https://kiwipy.readthedocs.io/en/latest/installation.html#building-from-source).

* https://github.com/aiidateam/kiwipy/pull/73
Adds benchmarks to the testing suite and a description on my finds in our documentation, [here](https://kiwipy.readthedocs.io/en/latest/performance.html).

I hope this helps to sufficiently addresses the points you've raised.

We look forward to the continuation of the review process.
