### 2020-07-20 aiidateam/kiwipy#75 @muhrin
Thanks for the report @dghoshal-lbl .  I'm just looking into how this could have come about.

Would you mind posting the results of `pip freeze` so I can see the versions of libraries that you have installed?
