### 2020-07-20 aiidateam/kiwipy#75 @muhrin
Could you also share your python version (`python --version`)?

The missing template (`AsyncContextManager`) has been available since python 3.5.4 and 3.6.2.  Is it possible that you have python < 3.6.2?
