### 2020-07-22 #2351 @muhrin
@danielskatz , just a quick update to say that I'm liasinging with @dghoshal-lbl to figure out the source of the exception they are experiencing.  The issue is being dealt with here https://github.com/aiidateam/kiwipy/issues/75.
