### 2020-07-27 aiidateam/kiwipy#75 @dghoshal-lbl
@muhrin You are correct. My python version was 3.6.1. After upgrading it, I was able to successfully run my test. Thanks!
