### 2020-07-27 #2351 @dghoshal-lbl
@muhrin I was able to run the test following our discussion in the github issue thread. All my questions have been addressed and I think the paper is good to go.

Just a couple of minor comments, which might be worth fixing:
i) It would be good to mention the required python version.
ii) For the test suite, it might be worth mentioning what all services should be running prior to running the tests. For example, the tests that use rabbitmq fail if the server is not running.
