### 2020-07-30 #2351 @muhrin
Thanks @dghoshal-lbl !

I've created a PR to address the final points, this should get merged soon: https://github.com/aiidateam/kiwipy/pull/76.
