### 2020-07-31 #2351 @danielskatz
@muhrin - please go ahead and merge your remaining PR.

Then, make sure the paper is complete and accurate, and

I can then move forward with accepting the submission, which include me proofreading the paper again.
