### 2020-08-03 ccfa794
Final changes responding to JOSS comments (#76)
 * Mention minimum Python patch versions
 * Make it explicit that RabbitMQ has to be running when testing if
   wanting to test the RMQ backend
