+++ kiwipy/version.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.6.0'
+__version__ = '0.6.1'
 
 __all__ = ('__version__',)
