### 2020-08-03 7b89a74
Add the logo to the `MANIFEST.in`
This way it will also show up on remote resources, for example on PyPI.
