### 2020-08-03 #2351 @muhrin
Thanks @danielskatz .  So, here it is:

Tagged version is: https://github.com/aiidateam/kiwipy/tree/v0.6.1

Zenodo archive can be found here: https://zenodo.org/record/3970461.

Metadata matches that of submission.

DOI is: 10.5281/zenodo.3970461

Let us know if you need anything else to proceed.
