### 2020-08-03 #2351 @danielskatz
@muhrin - There are 2 PRs for small issues in the paper and the bib

- https://github.com/aiidateam/kiwipy/pull/78
- https://github.com/aiidateam/kiwipy/pull/79

Please let me know when these are merged, or if you want to discuss them
