### 2020-08-03 0cff35e
removing 2 abstracts from bib (#78)
Removed abstracts from .bib file for JOSS publication.
