+++ docs/source/paper.bib
@@ -12,7 +12,6 @@ year = {2013}
 }
 
 @article{Talirz2020,
-abstract = {Materials Cloud is a platform designed to enable open and seamless sharing of resources for computational science, driven by applications in materials modelling. It hosts 1) archival and dissemination services for raw and curated data, together with their provenance graph, 2) modelling services and virtual machines, 3) tools for data analytics, and pre-/post-processing, and 4) educational materials. Data is citable and archived persistently, providing a comprehensive embodiment of the FAIR principles that extends to computational workflows. Materials Cloud leverages the AiiDA framework to record the provenance of entire simulation pipelines (calculations performed, codes used, data generated) in the form of graphs that allow to retrace and reproduce any computed result. When an AiiDA database is shared on Materials Cloud, peers can browse the interconnected record of simulations, download individual files or the full database, and start their research from the results of the original authors. The infrastructure is agnostic to the specific simulation codes used and can support diverse applications in computational science that transcend its initial materials domain.},
 archivePrefix = {arXiv},
 arxivId = {2003.12510},
 author = {Talirz, Leopold and Kumbhar, Snehal and Passaro, Elsa and Yakutovich, Aliaksandr V. and Granata, Valeria and Gargiulo, Fernando and Borelli, Marco and Uhrin, Martin and Huber, Sebastiaan P. and Zoupanos, Spyros and Adorf, Carl S. and Andersen, Casper W. and Sch{\"{u}}tt, Ole and Pignedoli, Carlo A. and Passerone, Daniele and VandeVondele, Joost and Schulthess, Thomas C. and Smit, Berend and Pizzi, Giovanni and Marzari, Nicola},
@@ -171,7 +170,6 @@ year = {2020}
 }
 
 @article{Huber2020,
-abstract = {The ever-growing availability of computing power and the sustained development of advanced computational methods have contributed much to recent scientific progress. These developments present new challenges driven by the sheer amount of calculations and data to manage. Next-generation exascale supercomputers will harden these challenges, such that automated and scalable solutions become crucial. In recent years, we have been developing AiiDA (http://www.aiida.net), a robust open-source high-throughput infrastructure addressing the challenges arising from the needs of automated workflow management and data provenance recording. Here, we introduce developments and capabilities required to reach sustained performance, with AiiDA supporting throughputs of tens of thousands processes/hour, while automatically preserving and storing the full data provenance in a relational database making it queryable and traversable, thus enabling high-performance data analytics. AiiDA's workflow language provides advanced automation, error handling features and a flexible plugin model to allow interfacing with any simulation software. The associated plugin registry enables seamless sharing of extensions, empowering a vibrant user community dedicated to making simulations more robust, user-friendly and reproducible.},
 archivePrefix = {arXiv},
 arxivId = {2003.12476},
 author = {Huber, Sebastiaan. P. and Zoupanos, Spyros and Uhrin, Martin and Talirz, Leopold and Kahle, Leonid and H{\"{a}}uselmann, Rico and Gresch, Dominik and M{\"{u}}ller, Tiziano and Yakutovich, Aliaksandr V. and Andersen, Casper W. and Ramirez, Francisco F. and Adorf, Carl S. and Gargiulo, Fernando and Kumbhar, Snehal and Passaro, Elsa and Johnston, Conrad and Merkys, Andrius and Cepellotti, Andrea and Mounet, Nicolas and Marzari, Nicola and Kozinsky, Boris and Pizzi, Giovanni},
