### 2020-08-03 4320fa0
Minor wording changes to the summary section of the paper (#79)
