### 2020-08-03 #2351 @danielskatz
@openjournals/dev - the DOI works but the PDF doesn't seem to load - can you check?
